% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cranstats-package.R
\docType{package}
\name{cranstats-package}
\alias{cranstats}
\alias{cranstats-package}
\title{cranstats: A Collection of Functions to Obtain Package Data from CRAN.}
\description{
A simple collection of functions for obtaining packages data from CRAN. These functions were inspired by a post on stack overflow and I decided to put them together for future reference.
}
\seealso{
Useful links:
\itemize{
  \item \url{}
  \item Report bugs at \url{}
}

}
\keyword{internal}
